+++ src/store/auth0/index.js
@@ -222,18 +222,27 @@ const actions = {
   async authenticate({ dispatch, commit, getters }) {
     if (!auth0Client) await dispatch('initAuth0')
 
-    if (window.location.search.includes('code=')) {
-      await auth0Client.handleRedirectCallback()
+    if (window.location?.search?.includes('code=')) {
+      try {
+        await auth0Client.handleRedirectCallback()
+      } catch {
+        // Catches repeated calls to the handleRedirectCallback...
+        // The first time handleRedirectCallback is called, it removes the challenge
+        // from the window search... when another call is made, it looks for the challenge
+        // which was removed, halting further execution and navigation. This allows the method to proceed
+        // in cases of an invalid state
+        // Don't do anything with this though.
+      }
     }
 
     const isAuthenticated = await auth0Client.isAuthenticated()
     commit('isAuthenticated', isAuthenticated)
 
-    if (window.location.pathname && !getters['redirectRoute']) {
+    if (window.location?.pathname && !getters['redirectRoute']) {
       dispatch('setRedirectRoute', window.location.pathname)
     }
 
-    if (window.location.search.includes('invitation_id=')) {
+    if (window.location?.search?.includes('invitation_id=')) {
       const invitationID = window.location.search.split('=')
       localStorage.setItem('invitationId', invitationID[1])
     }