+++ Changelog.md
@@ -5,6 +5,15 @@
 ### Features and Improvements
 
 - Use labels from `flow.run_config` if present [#309](https://github.com/PrefectHQ/ui/pull/309)
+
+### Bugfixes
+
+- None
+
+## 2020-10-09
+
+### Features and Improvements
+
 - Add section to flow details for `flow.run_config` if not null, otherwise display `flow.environment` [#307](https://github.com/PrefectHQ/ui/pull/307)
 - Enable restart from cancelled and update restart from failed [#234](https://github.com/PrefectHQ/ui/pull/234)
 - Add unit tests for the authNavGuard middleware [#266](https://github.com/PrefectHQ/ui/pull/266)
@@ -19,6 +28,7 @@
 - Fix bug with unknown invitation ID in accept invitation [#306](https://github.com/PrefectHQ/ui/pull/306)
 - Fix race condition with onboard/welcome flow [#304](https://github.com/PrefectHQ/ui/pull/304)
 - Fix version display in flow page version dropdown - [#305](https://github.com/PrefectHQ/ui/pull/305)
+- Invalid state hotfix - [#308](https://github.com/PrefectHQ/ui/pull/308)
 
 ## 2020-10-05